+++ crates/nesmc-emu/src/nes_machine/bus/input.rs
@@ -4,9 +4,9 @@ use super::Device;
 pub struct Input;
 
 impl Device for Input {
-    fn read(&self, _addr: usize) -> u8 {
+    fn read(&self, _addr: u16) -> u8 {
         0
     }
 
-    fn write(&mut self, _addr: usize, _value: u8) {}
+    fn write(&mut self, _addr: u16, _value: u8) {}
 }
