+++ crates/nesmc-emu/src/nes_machine/bus/ppu.rs
@@ -4,12 +4,12 @@ use super::Device;
 pub struct PPU;
 
 impl Device for PPU {
-    fn read(&self, _addr: usize) -> u8 {
+    fn read(&self, _addr: u16) -> u8 {
         //let addr = addr % 0x8;
         0
     }
 
-    fn write(&mut self, _addr: usize, _value: u8) {
+    fn write(&mut self, _addr: u16, _value: u8) {
         //let addr = addr % 0x8;
     }
 }
